# [no-release-notes] Docstrings for chunker