Logical refactor that makes it easier to add documentation for
chunker.AdvanceTo.