@max-hoffman 2022-06-01
benchmarks after
```
BenchmarkMapUpdate/benchmark_maps_10k/benchmark_new_format_writes
BenchmarkMapUpdate/benchmark_maps_10k/benchmark_new_format_writes-12         	   19617	     59245 ns/op	   19347 B/op	      54 allocs/op
BenchmarkMapUpdate/benchmark_maps_10k/benchmark_old_format_writes
BenchmarkMapUpdate/benchmark_maps_10k/benchmark_old_format_writes-12         	    1969	    637116 ns/op	  481815 B/op	    3173 allocs/op
BenchmarkMapUpdate/benchmark_maps_100k
BenchmarkMapUpdate/benchmark_maps_100k/benchmark_new_format_writes
BenchmarkMapUpdate/benchmark_maps_100k/benchmark_new_format_writes-12        	   19452	     62476 ns/op	   20893 B/op	      54 allocs/op
BenchmarkMapUpdate/benchmark_maps_100k/benchmark_old_format_writes
BenchmarkMapUpdate/benchmark_maps_100k/benchmark_old_format_writes-12        	    1320	   1015432 ns/op	  713799 B/op	    4907 allocs/op
BenchmarkMapUpdate/benchmark_maps_1M
BenchmarkMapUpdate/benchmark_maps_1M/benchmark_new_format_writes
BenchmarkMapUpdate/benchmark_maps_1M/benchmark_new_format_writes-12          	   13495	     89753 ns/op	   28831 B/op	      54 allocs/op
BenchmarkMapUpdate/benchmark_maps_1M/benchmark_old_format_writes
BenchmarkMapUpdate/benchmark_maps_1M/benchmark_old_format_writes-12          	    1263	   1028665 ns/op	  819051 B/op	    5157 allocs/op
BenchmarkProllySmallWrites
BenchmarkProllySmallWrites/benchmark_new_format_writes
BenchmarkProllySmallWrites/benchmark_new_format_writes-12                    	   21049	     56210 ns/op	   19158 B/op	      54 allocs/op
BenchmarkTypesSmallWrites
BenchmarkTypesSmallWrites/benchmark_old_format_writes
BenchmarkTypesSmallWrites/benchmark_old_format_writes-12                     	    1732	    594336 ns/op	  486973 B/op	    3219 allocs/op
BenchmarkProllyMediumWrites
BenchmarkProllyMediumWrites/benchmark_new_format_writes
BenchmarkProllyMediumWrites/benchmark_new_format_writes-12                   	   19257	     61866 ns/op	   20920 B/op	      54 allocs/op
BenchmarkTypesMediumWrites
BenchmarkTypesMediumWrites/benchmark_old_format_writes
BenchmarkTypesMediumWrites/benchmark_old_format_writes-12                    	    1269	    905638 ns/op	  711083 B/op	    4872 allocs/op
BenchmarkProllyLargeWrites
BenchmarkProllyLargeWrites/benchmark_new_format_writes
BenchmarkProllyLargeWrites/benchmark_new_format_writes-12                    	   13502	     87961 ns/op	   28918 B/op	      54 allocs/op
```
benchmarks before:
```
BenchmarkMapUpdate/benchmark_maps_10k
BenchmarkMapUpdate/benchmark_maps_10k/benchmark_new_format_writes
BenchmarkMapUpdate/benchmark_maps_10k/benchmark_new_format_writes-12         	   20390	     58255 ns/op	   19321 B/op	      54 allocs/op
BenchmarkMapUpdate/benchmark_maps_10k/benchmark_old_format_writes
BenchmarkMapUpdate/benchmark_maps_10k/benchmark_old_format_writes-12         	    1990	    613679 ns/op	  482894 B/op	    3184 allocs/op
BenchmarkMapUpdate/benchmark_maps_100k
BenchmarkMapUpdate/benchmark_maps_100k/benchmark_new_format_writes
BenchmarkMapUpdate/benchmark_maps_100k/benchmark_new_format_writes-12        	   19617	     62682 ns/op	   20897 B/op	      54 allocs/op
BenchmarkMapUpdate/benchmark_maps_100k/benchmark_old_format_writes
BenchmarkMapUpdate/benchmark_maps_100k/benchmark_old_format_writes-12        	    1310	   1005681 ns/op	  713477 B/op	    4872 allocs/op
BenchmarkMapUpdate/benchmark_maps_1M
BenchmarkMapUpdate/benchmark_maps_1M/benchmark_new_format_writes
BenchmarkMapUpdate/benchmark_maps_1M/benchmark_new_format_writes-12          	   13424	     88028 ns/op	   28905 B/op	      54 allocs/op
BenchmarkMapUpdate/benchmark_maps_1M/benchmark_old_format_writes
BenchmarkMapUpdate/benchmark_maps_1M/benchmark_old_format_writes-12          	    1167	    983159 ns/op	  812032 B/op	    5046 allocs/op
BenchmarkProllySmallWrites
BenchmarkProllySmallWrites/benchmark_new_format_writes
BenchmarkProllySmallWrites/benchmark_new_format_writes-12                    	   21463	     55591 ns/op	   19042 B/op	      54 allocs/op
BenchmarkTypesSmallWrites
BenchmarkTypesSmallWrites/benchmark_old_format_writes
BenchmarkTypesSmallWrites/benchmark_old_format_writes-12                     	    2000	    595175 ns/op	  482842 B/op	    3172 allocs/op
BenchmarkProllyMediumWrites
BenchmarkProllyMediumWrites/benchmark_new_format_writes
BenchmarkProllyMediumWrites/benchmark_new_format_writes-12                   	   19276	     61658 ns/op	   20923 B/op	      54 allocs/op
BenchmarkTypesMediumWrites
BenchmarkTypesMediumWrites/benchmark_old_format_writes
BenchmarkTypesMediumWrites/benchmark_old_format_writes-12                    	    1173	    965731 ns/op	  717874 B/op	    4952 allocs/op
BenchmarkProllyLargeWrites
BenchmarkProllyLargeWrites/benchmark_new_format_writes
BenchmarkProllyLargeWrites/benchmark_new_format_writes-12                    	   13621	     87070 ns/op	   28886 B/op	      54 allocs/op
```